feat(modem): Add modem fw version metric
It is useful to have the modem firmware in a device attribute since
that firmware is independently flashable from the main application
and there's a requirement for it to be compatible. Upon boot, the
modem firmware version is fetched and a string metric is updated
with this value. The first heartbeat will have this value, and it will
be viewable as an attribute.

Tested by capturing a heartbeat before and after an update to the
modem firmware to see the reported version change.

Before:
<img width="567" alt="before" src="https://github.com/memfault/sdk-nrf/assets/41022382/f2f92d56-29e7-424c-8eca-8ccf313ef3f8">

After:
<img width="576" alt="after" src="https://github.com/memfault/sdk-nrf/assets/41022382/5b89dfb8-be98-4649-bc10-c91da856de3e">

Changes to the Kconfig with a new dependency were tested by removing
`CONFIG_MODEM_INFO` from both `sdk/embedded/examples/nrf-connect-sdk/nrf9160/memfault_demo_app/CMakeLists.txt`
and `sdk-nrf/modules/memfault-firmware-sdk/Kconfig`. The following
warning is generated:

<img width="814" alt="Screenshot 2023-08-28 at 11 37 13 AM" src="https://github.com/memfault/sdk-nrf/assets/41022382/76d17563-1e5b-4a4d-a264-83ee244e98e4">

Warning goes away after re-adding `CONFIG_MODEM_INFO`.