tfm: Handle NS busfaults with Zephyr fault infrastructure
  ### Background Context

In NCS v2.1.0, Secure Partition Manager (SPM) was deprecated in favor
of TFM for the trusted execution environment on the nRF53/91
series. This cutover resulted in a change in behavior causing all
BusFaults to get forwarded to the TFM fault handlers _even_ when they
are caused by non-secure firmware and Zephyrs fault handler never gets
called.

This is problematic for several reasons:

 - Slows down local debug because fault information no longer gets
   printed in Zephyr shell.
 - Customers debugging failures remotely in the field (using tools
   like Memfault [^1]) no longer get information about any busfaults
   (caused by common programmatic errors like a null pointer
   dereference).

[^1]: https://developer.nordicsemi.com/nRF_Connect_SDK/doc/latest/nrf/libraries/debug/memfault_ncs.html

This PR restores SPM-like functionality by calling the Zephyr fault
handler when a busfault is triggered by non-secure code.

  ### Summary of Changes

 - enabled capturing of exception info when TFM fault handlers are
   invoked (required `-DTFM_EXCEPTION_INFO_DUMP=True`)
 - Added a small change[^2] to sdk-trusted-firmware-m to expose exception info via API
 - Implemented custom platform reset handlers
   PLATFORM_DEFAULT_SYSTEM_RESET_HALT (instead of using the default
   ones from [platform/ext/common/tfm_hal_reset_halt.c [^3]

The function of most interest to review is `tfm_hal_system_reset()`
which takes advantage of a few properities related to ARMv8-M
exception handling (see the Armv8-M Architecture Reference Manual
[^4]) for more details. Notably,

 - BusFault's are not banked between NS/S and by default will generate
   a secure exception.
 - Any attempt to access secure memory will trigger a SecureFault
 - When a Secure exception is executed we can use the info in
   EXC_RETURN to determine whether the fault originated from the
   Secure or Non-Secure side.

Piecing these properties together -- from the SecureFault and BusFault
handlers we check to see if the fault originated from the non-secure
side and iff that is the case, invoke the non-secure Zephyr fault
handler to save off exception info. Note these handlers are at
configurable interrupt priority and AIRCR.PRIS is set -- so any error
from the Zephyr fault handler will trap back into the secure HardFault
handler and the device will just be reset.

[^2]: https://github.com/nrfconnect/sdk-trusted-firmware-m/compare/master...memfault:sdk-trusted-firmware-m:ccoleman/exc_info?expand=1
[^3]: https://github.com/nrfconnect/sdk-trusted-firmware-m/blob/master/platform/ext/common/tfm_hal_reset_halt.c
[^4]: https://developer.arm.com/documentation/ddi0553/bw/?lang=en

  ### Test Plan

 - Tested against v2.4 with the above patch applied to sdk-nrf and
   trusted-firmware-m
 - Instrument code to trigger a null pointer dereference and observe
   that the Zephyr fault handler gets invoked

```
  volatile uint32_t *null_ptr =(void *)0x0;
  *null_ptr = 0xbadcafe;
```
 - Flashed `nrf/samples/debug/memfault` sample and confirmed fault
   paths which previously did not get captured, now invoke the Zephyr
   handler:
   - `mflt test busfault`
   - `mflt test double_free`