Use new metrics APIs (#13)
We have a new metrics API, so I bumped the memfault version and removed
most references to the old API.
The only references I left were in the collection of thread metrics
[here](https://github.com/nrfconnect/sdk-nrf/blob/dc672091f5c7dcd2746856e02ac0ba1cf15837d6/modules/memfault-firmware-sdk/memfault_ncs_metrics.c#L74)
and in asset tracker debug module unit tests
[here](https://github.com/nrfconnect/sdk-nrf/blob/dc672091f5c7dcd2746856e02ac0ba1cf15837d6/applications/asset_tracker_v2/tests/debug_module/src/debug_module_test.c#L116).
The module that collects thread metrics needs to share references to the
metrics with
other modules, so storing the key and using the original metrics API
makes sense.
The unit tests will need a reference to the underlying function called
for mocking so we can't use
the `MEMFAULT_METRIC_*` macro in this case.

Also fixed some incorrect formatting that slipped through.

Built with the asset tracker app and confirmed a successful heartbeat
pushed to Memfault.

----

Resolves: MFLT-12884

---------

Co-authored-by: Eric Johnson <eric.j.johnson22@gmail.com>